backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic[email]
email-validator
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==1.7.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
itsdangerous==2.2.0
authlib==1.6.5
google-auth
httpx
requests
